requirements.txt
celery[redis]
psycopg2-binary
requests
python-dateutil
redis